Synchronize the code

+++ consensus/poc-consensus/src/main/java/io/nuls/poc/model/dto/input/CreateAgentDTO.java
@@ -11,7 +11,7 @@ public class CreateAgentDTO {
     private String agentAddress;
     private String packingAddress;
     private String rewardAddress;
-    private double commissionRate;
+    private byte commissionRate;
     private String deposit;
     private String password;
 
@@ -47,11 +47,11 @@ public class CreateAgentDTO {
         this.rewardAddress = rewardAddress;
     }
 
-    public double getCommissionRate() {
+    public byte getCommissionRate() {
         return commissionRate;
     }
 
-    public void setCommissionRate(double commissionRate) {
+    public void setCommissionRate(byte commissionRate) {
         this.commissionRate = commissionRate;
     }
